Let event callbacks report an error or not
When polling a FD yields an error, for whatever reason, the application
cannot communicate this to udaemon. This makes it difficult to recover
from such errors as it might be that poll keeps reporting the same
status for a faulty FD over and over again. By returning `RES_ERROR`
from an event handler, udaemon can simply close the offending FD and
continue with other tasks.